Switch to routing table style
